pyproject.toml
[tool.poetry]
name = "gestion-de-stock"
version = "0.1.0"
description = "Aplicación para gestionar inventarios y productos."
authors = ["Mauro <mausrosales@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"
werkzeug = "^3.1.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
